# docs/swagger.yaml
definitions:
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  model.Comment:
    properties:
      comment_message:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo:
        $ref: '#/definitions/model.Photo'
      photo_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.Photo:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photo_caption:
        example: photo caption
        type: string
      photo_title:
        example: photo title
        type: string
      photo_url:
        example: photo url
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.SocialMedia:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
...
  model.SwagSocialMedia:
    properties:
      name:
        example: My Name
        type: string
      social_media_url:
        example: https://socialurl.com
        type: string
    type: object
  model.SwagUserLogin:
    properties:
      user_email:
        example: my_email@address.com
        type: string
      user_password:
        example: my_password
        type: string
    type: object
  model.SwagUserRegister:
    properties:
      user_age:
        example: 15
        type: integer
      user_email:
        example: my_email@address.com
        type: string
      user_password:
        example: my_password
        type: string
      username:
        example: my_username
        type: string
    type: object
  model.User:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photos:
        items:
          $ref: '#/definitions/model.Photo'
        type: array
      updated_at:
        type: string
      user_age:
        type: integer
      user_email:
        type: string
      user_password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /social_medias:
    get:
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SocialMedia'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Return all social medias
      tags:
      - Social Media
    post:
      consumes:
      - application/json
      description: |-
        Specified user was taken from the token. Post a json data to add social media.
        Validates: (Name, Social Media URL)
        Error: No duplicates
      parameters:
      - description: '-'
        in: body
        name: socialMedia
        required: true
        schema:
          $ref: '#/definitions/model.SwagSocialMedia'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Add social media to a user
      tags:
      - Social Media
  /social_medias/{id}:
    delete:
      description: |-
        Return forbidden if the user is not the owner of the social media
        Validates: (Name, Social Media URL)
      parameters:
      - description: '-'
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Delete certain social media that is specified by the id
      tags:
      - Social Media
    get:
      consumes:
      - application/json
      description: Input valid social media id into parameter
      parameters:
      - description: Social Media ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SocialMedia'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Get social media by social media id
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: |-
        Return forbidden if the user is not the owner of the social media
        Validates: (Name, Social Media URL)
      parameters:
      - description: Social Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: socialMedia
        required: true
        schema:
          $ref: '#/definitions/model.SwagSocialMedia'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Update certain social media that is specified by the id
      tags:
      - Social Media
  /users/login:
    post:
      consumes:
      - application/json
      description: |-
        post a json data to login
        Validates: (Email, Password)
      parameters:
      - description: Email must be valid, Password longer than 5 characters
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.SwagUserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: string
            - properties:
                status:
                  type: integer
                token:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Log certain user in
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: |-
        post a json data to register
        Validates: (Username, Email, Password, Age[13 >])
      parameters:
      - description: user
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/model.SwagUserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Register a user
      tags:
      - User
swagger: "2.0"